Add contact_details and degrees to end-to-end spec
The candidate_completing_an_application_spec spec now runs through the
entire application form.

To make that happen this commit:

- extracts the form-filling logic for contact details and degrees into
helper modules, in line with what we did for personal details
- includes those in the end to end spec
- fixes a bug revealed by that end to end spec, which is that completing
contact details takes you to the summary page, not to the "Add a degree"
page
- corrects copy-pasta in the Degree spec (we visited
contact_details/new, not degrees/new)
