+++ spec/system/candidate_entering_contact_details_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 
 describe 'A candidate entering contact details' do
-  include TestHelpers::PersonalDetails
+  include TestHelpers::ContactDetails
 
   context 'when details are correct' do
     before do
@@ -11,12 +11,10 @@ describe 'A candidate entering contact details' do
       click_on t('application_form.save_and_continue')
     end
 
-    describe 'check_your_answers page' do
-      it 'contains phone number' do
-        expect(page).to have_content('Phone number')
-        expect(page).to have_content('Email address')
-        expect(page).to have_content('Postal address')
-      end
+    it 'sees a summary of those details' do
+      visit '/check_your_answers'
+
+      expect(page).to have_content('Phone number 1234567890')
     end
 
     context 'and wishes to amend their details' do
@@ -40,19 +38,5 @@ describe 'A candidate entering contact details' do
       click_on 'Enter your phone number'
       expect(page).to have_selector('#contact_details_phone_number:focus')
     end
-  end
-
-private
-
-  def fill_in_contact_details
-    details = {
-      phone_number: '1234567890',
-      email_address: 'john.doe@example.com',
-      address: 'Westminster, London SW1P 1QW'
-    }
-
-    fill_in t('application_form.contact_details_section.phone_number.label'), with: details[:phone_number]
-    fill_in t('application_form.contact_details_section.email_address.label'), with: details[:email_address]
-    fill_in t('application_form.contact_details_section.address.label'), with: details[:address]
   end
 end
