+++ spec/system/candidate_entering_degree_spec.rb
@@ -1,6 +1,8 @@
 require 'rails_helper'
 
 describe 'A candidate adding a Degree' do
+  include TestHelpers::DegreeDetails
+
   context 'who successfully enters their details' do
     before do
       visit '/degrees/new'
@@ -26,7 +28,7 @@ describe 'A candidate adding a Degree' do
 
   context 'who leaves out a required field' do
     before do
-      visit '/contact_details/new'
+      visit '/degrees/new'
     end
 
     it 'sees an error summary' do
@@ -34,23 +36,5 @@ describe 'A candidate adding a Degree' do
 
       expect(page).to have_content('There is a problem')
     end
-  end
-
-private
-
-  def fill_in_degree_details
-    details = {
-      type: 'BA',
-      subject: 'Philosophy',
-      institution: 'University of London',
-      class: 'first',
-      year: 2000
-    }
-
-    fill_in t('application_form.degree_details_section.type.label'), with: details[:type]
-    fill_in t('application_form.degree_details_section.subject.label'), with: details[:subject]
-    fill_in t('application_form.degree_details_section.institution.label'), with: details[:institution]
-    fill_in t('application_form.degree_details_section.class.label'), with: details[:class]
-    fill_in t('application_form.degree_details_section.year.label'), with: details[:year]
   end
 end
